backend/requirements.txt
fastapi==0.117.1
langchain==0.3.27
langchain_core==0.3.76
langgraph==0.6.7
pydantic==2.11.9
python-dotenv==1.1.1
typing_extensions==4.15.0
uvicorn==0.35.0
langchain_community
langchain-google-genai
sentence-transformers
chromadb
numpy

